# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LanguageCode = typeof LanguageCode[keyof typeof LanguageCode];


export const LanguageCode = {
  en: 'en',
  ja: 'ja',
  zh: 'zh',
  es: 'es',
} as const;

export interface AnalyzeScenePayload {
  /** Base64-encoded JPEG/PNG image data (no data URI prefix) */
  imageBase64: string;
  language: LanguageCode;
}

/**
 * Normalized coordinates (0 to 1) relative to image width/height
 */
export interface BoundingBox {
  x: number;
  y: number;
  width: number;
  height: number;
}

export interface DetectedObject {
  id: string;
  /** The object's name in English */
  englishLabel: string;
  /** The object's name translated into the target language */
  word: string;
  /** Romanized/phonetic reading for non-Latin scripts, null otherwise */
  romanization: string | null;
  /** Korean Hangul transcription of how the word is pronounced */
  koreanPronunciation: string | null;
  boundingBox: BoundingBox;
}

export interface AnalyzeSceneResult {
  objects: DetectedObject[];
}

export interface SynthesizeSpeechBody {
  text: string;
  language: LanguageCode;
}

export interface SynthesizeSpeechResponse {
  audioBase64: string;
  format: string;
}