# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vocab
    description: Visual vocabulary detection and pronunciation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /vocab/analyze:
    post:
      operationId: analyzeScene
      tags: [vocab]
      summary: Analyze a photo and return detected objects with vocabulary labels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeScenePayload"
      responses:
        "200":
          description: Detected objects with translated vocabulary and bounding boxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeSceneResult"
  /vocab/speech:
    post:
      operationId: synthesizeVocabSpeech
      tags: [vocab]
      summary: Synthesize spoken pronunciation for a vocabulary word
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SynthesizeSpeechBody"
      responses:
        "200":
          description: Base64-encoded audio of the spoken word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SynthesizeSpeechResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LanguageCode:
      type: string
      enum: [en, ja, zh, es]
    AnalyzeScenePayload:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded JPEG/PNG image data (no data URI prefix)
        language:
          $ref: "#/components/schemas/LanguageCode"
      required:
        - imageBase64
        - language
    DetectedObject:
      type: object
      properties:
        id:
          type: string
        englishLabel:
          type: string
          description: The object's name in English
        word:
          type: string
          description: The object's name translated into the target language
        romanization:
          type: string
          nullable: true
          description: Romanized/phonetic reading for non-Latin scripts, null otherwise
        koreanPronunciation:
          type: string
          nullable: true
          description: Korean Hangul transcription of how the word is pronounced
        boundingBox:
          $ref: "#/components/schemas/BoundingBox"
      required:
        - id
        - englishLabel
        - word
        - romanization
        - koreanPronunciation
        - boundingBox
    BoundingBox:
      type: object
      description: Normalized coordinates (0 to 1) relative to image width/height
      properties:
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
      required:
        - x
        - y
        - width
        - height
    AnalyzeSceneResult:
      type: object
      properties:
        objects:
          type: array
          items:
            $ref: "#/components/schemas/DetectedObject"
      required:
        - objects
    SynthesizeSpeechBody:
      type: object
      properties:
        text:
          type: string
        language:
          $ref: "#/components/schemas/LanguageCode"
      required:
        - text
        - language
    SynthesizeSpeechResponse:
      type: object
      properties:
        audioBase64:
          type: string
        format:
          type: string
      required:
        - audioBase64
        - format