# lib/api-zod/src/generated/types/detectedObject.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoundingBox } from './boundingBox';

export interface DetectedObject {
  id: string;
  /** The object's name in English */
  englishLabel: string;
  /** The object's name translated into the target language */
  word: string;
  /** Romanized/phonetic reading for non-Latin scripts, null otherwise */
  romanization: string | null;
  /** Korean Hangul transcription of how the word is pronounced */
  koreanPronunciation: string | null;
  boundingBox: BoundingBox;
}